Payment page and credit card integration added

+++ ECommerceApp/ECommerceApp.Web/Models/Cart.cs
@@ -26,7 +26,7 @@ namespace ECommerceApp.Web.Models
         {
             Items.RemoveAll(i => i.Product.Id == product.Id);
         }
-        public decimal CalculateTotal()
+        public double CalculateTotal()
         {
             return Items.Sum(i => i.Product.Price * i.Quantity);
         }
